python bindings: use variable number of arguments in overloaded methods
Even though overloaded methods typically have the same number
of arguments encoded in the function name (and therefore also
the same total number of arguments), it is in theory possible
for them to have a different number of arguments.
This can happen in particular if the only argument is an isl_ctx.
Support for handling this still needs to be added
to the bindings generator, but this isl_ctx would get dropped
from the Python method, while any other overload would
preserve its single argument.

Using a variable number of arguments in overloaded methods
also makes their handling more similar to constructors,
allowing some of the code to be reused.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>