python bindings: support automatic conversions in overloaded methods
The commit message of isl-0.21-8-g7fe7f37942 (C++ bindings:
automatically convert long arguments to isl::val,
Thu Dec 27 14:45:28 2018 +0100) claims that automatic conversion
is not needed for the Python bindings, because any argument is
already automatically converted to the expected type and
because the context is implicit in the Python bindings.
However, this only applies to regular methods and
not to overloaded methods.

Also perform the automatic conversion for overloaded methods.
As an example, this allows the add_named_tuple method
of isl.space to be called with a string name instead of an isl.id name.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>