cpp_generator::class_printer::print_method_variants: pass in method name
The method name is already known in
the cpp_generator::class_printer::print_methods caller,
so it might as well be passed down.
This will make it easier to handle methods that appear
in a given class, but that are derived from a function declaration
for a different isl type.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>