configure interface subdirectory separately
The process of compiling extract_interface is quite different
from that of compiling isl itself.  The required configuration
is also essentially distinct.  Split the configuration by
configuring the interface subdirectory separately.
This will be especially useful when adding support
for cross-compilation with bindings generation.

Note that this requires autogen.sh and configure to be rerun.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>