disable shared libraries when building extract_interface
There is really no point in using shared libraries
for building extract_interface as no libraries
need to get built anyway.
However, this changes the behavior of libtool.
In particular, if building of shared libraries is not disabled,
then libtool gets confused when compiling with the build compiler
instead of the host compiler.
It is still useful to use libtool however, as it handles
-R arguments in a uniform way.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>