interface: Method: keep track of method name
Keep track of the method name inside the Method object
such that it does not need to be passed along separately.
In fact, provide a constructor for setting
a default name such that the name only needs to be specified
when it is different from the default.

Any renaming of the method name is also taken care of
by the Method constructor.
Since cpp_generator::rename_method is only used
from this constructor and does not need any
information from the cpp_generator object,
it is turned into a static function.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>