interface: introduce cpp_generator::class_printer
The original implementation of the C++ bindings generator
had quite a bit of duplicate logic for figuring out
which methods to print for both declarations and implementations,
with only the actual method printing being different.
The method printing methods were later templatized
to remove some of this duplication.
However, this is actually more appropriately handled
using polymorphism, with a shared implementation
for deciding which methods to print and
specializations for actually printing the methods.
The template methods are therefore replaced by methods
is a new class printing class, with specializations
for printing declarations and implementations.
This has the added advantage that information that
is shared by all (or at least most) printing methods
does not need to be passed along every time,
but can instead be stored in the class printer.
Most notably, this is the description of the class methods,
but the output stream can also be stored in the class printer.
The C++ name of the class is also stored in the class printer
because it is computed in the same way in several places and
it will be further reused in upcoming commits.

Note that the decl_printer is currently constructed four times,
while the impl_printer is currently constructed three times.
In the end, only cpp_generator::print_class and
cpp_generator::print_class_impl will construct a printer,
but these are not currently passed to the other methods,
while they do need to call print_method or print_method_variants.
The extra constructions will be removed in upcoming commits.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>