rename cpp_generator to plain_cpp_generator
An upcoming commit will add support for generating a templated
C++ interface.  This will reuse the generic part
of the current (plain) C++ interface generator.
The entire generator is first renamed to plain_cpp_generator and
the generic pieces will be moved back to cpp_generator
in a later commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>