class_printer::print_method_header: extract out print_full_method_header
When support for generating a templated C++ interface gets added,
it will be useful to be able to print extra elements
before the terminating semicolon of the method declaration.
Extract out a print_full_method_header that also prints
the semicolon and that will end up in the plain_cpp_generator,
while class_printer::print_method_header will remain
in the (then) generic cpp_generator.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>