interface: introduce Method::{c_,}num_params
This allows EnumMethod to override the num_params method to remove
the final argument, obviating the need for a print_method_header
variant taking the number of arguments.

The c_num_params method is needed to make the distinction between
the number of arguments in the method declaration and
the number of arguments in the call of the corresponding C function.
This method is only used by cpp_generator::print_method,
where its result is currently always equal to the result of num_params,
but this will change when EnumMethod definitions
also get printed using cpp_generator::print_method in the next commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>