export cond to the templated C++ interface
Note that it is plausible to allow the signatures of the
second and third arguments to be more general in the future
(e.g. they could be of a "multi" type). However, since
the only *_cond function takes isl_pw_aff arguments,
which are necessarily anonymous functions, it is currently
sufficient to require all arguments to be anonymous functions.

Signed-off-by: Tianjiao Sun <tianjiao.sun@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven@cerebras.net>