python_generator::print_method_call: recompute drop_user
In the next commit, support will be added for multiple callbacks.
This requires python_generator::print_method_call to be aware
of the number of user pointer arguments that have been skipped
so far, not only the total number of user pointer arguments.
This means python_generator::print_method_call needs
to recompute drop_user anyway.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>