export isl_*_gist_params to templated interface
These complement the already exported isl_*_gist and isl_*_gist_domain.
Note that a function such as isl_set_gist_params can also
be applied to a pair of parametric sets, but it is more
appropriate to call isl_set_gist in that case.
"gist_params" is therefore not exported as a "bin_params" method.
Note that "intersect_params" is also not exported as a "bin_params" method.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>