python interface: convert first argument of overloaded methods to correct types
The reasoning of isl-0.12.1-7-g3a637dd3b2 (python interface: also convert
first argument to correct type, Fri Sep 27 18:24:54 2013 +0200).
also applies to overloaded methods.  That is, just as in the case
of non-overloaded methods, invoking an overloaded method on an object
of a class that does not override the method defined by a base class
ends up calling the base class method on a reference to a C object
that corresponds to the derived class.
The object on which the overloaded method is invoked therefore also needs
to be explicitly converted to the base class.
This was missing from isl-0.15-213-g1d1813c0fd (python interface: add support
for overloaded methods, Fri Nov 20 16:18:01 2015 +0100).

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>